Improvements to exception handling to transport structured tracebacks.
This code is still fairly hackish, but we're starting to get there.
We still need to improve the api, because right now runlines() does
way too much, and I had to set the exception state in a temporary
private variable.  But client-side things are working better, so we
can continue fixing the kernel without bottlenecking Evan.