Propertly keep track of prompt position
Since the QTextEdit object has a maximumBlockCount constraint, lines of
text can get removed from the beginning of the document at any time we
append to the document.

It is important to keep proper track of the beginning and end position
of the prompt, also when text is inserted or removed. Instead of
manaully tracking this position by computing offsets when doing text
operations, this commit uses QTextCursor objects to do this.

This commit fixes a bug where self._append_before_prompt_pos is
incorrectly computed, due to text being removed as the prompt is being
appended to the document.

fixes #94, #107